feat: Captchas!!
We now got captchas, Certain routes can have Captchas enabled, We also allow searching for certain things inside the body for if the captcha should trigger.

We are using Cloudflares turnstile. We also got the option for Action though not implemented it much since it needs to be done client side before server side can deal with it.